[T-1] First commit, basics done.
### Summary

This commit marks the first step of the project. At this stage, the project is functional. An existing implementation of UVM for C# NuGet generation has been developed to verify that UVM includes all necessary functions and that they work as expected.

We have successfully generated all packages in the correct order with custom upgrade rules.

We hope that this library will be useful for package generation across multiple programming languages.

### Major

- Add : Implementation of the UVM.Interface, which provides external users with a set of properties and functions to implement, enabling UVM to work with their project.
- Add : Implementation of UVM.Logging, an internal logging tool that captures logs for better support and troubleshooting.
- Add : Implementation of UVM.Engine, a method library that includes tools for dependency tree generation, upgrading, dumping, and packing.
- Add : Implementation of UVM.Service, a library providing git wrapper tools to simplify the usage of the UVM.Engine.
- Add : Implementation of UVM.SDK, a NuGet package that includes UVM.Interface, UVM.Engine, and UVM.Service to reduce the number of dependencies required for UVM users.
- Add : UVM.Common : 1.0.0.0
- Add : UVM.Logging : 1.0.0.0
- Add : UVM.Engine : 1.0.0.0
- Add : UVM.Service : 1.0.0.0
- Add : UVM.SDK : 1.0.0.0

### Minor

...

###  Patch

...

### Note

We are aware that UVM is functional at this stage but lacks testing and CI/CD integration. Our next focus will be on addressing those areas before adding any new features to UVM.

Additionally, we are working on another project that will embed UVM and include more development tools. More details on that will be shared later.

Best regards,
The Team